% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ahp_functions.R
\name{AhpMatrix}
\alias{AhpMatrix}
\title{Create the AHP preference matrix from a dataframe containing
the pairwiswe preferences.}
\usage{
AhpMatrix(preferenceCombinations)
}
\arguments{
\item{preferenceCombinations}{a data.frame containing category or alternative
A in the first column, B in the second colum, and the preference in the third column.}
}
\value{
an AHP preference matrix
}
\description{
Create the AHP preference matrix from a dataframe containing
the pairwiswe preferences.
}

