% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analyze.R
\name{Analyse}
\alias{Analyse}
\title{Calculates the contribution of criteria preferences
to the overall decision.}
\usage{
Analyse(tr)
}
\arguments{
\item{tr}{the calculated AHP data.tree}
}
\value{
a data.frame containing the contribution of each criteria
}
\description{
Calculates the contribution of criteria preferences
to the overall decision.
}

