% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/alternative.R
\name{as.data.frame.AlternativesList}
\alias{as.data.frame.AlternativesList}
\title{Convert an AlternativesList object into a dataframe.
This is particularly useful for printing the results of an AHP study.}
\usage{
\method{as.data.frame}{AlternativesList}(alternatives, row.names = NULL,
  optional = FALSE, ...)
}
\description{
Convert an AlternativesList object into a dataframe.
This is particularly useful for printing the results of an AHP study.
}

