% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calculate.R
\name{Calculate}
\alias{Calculate}
\title{Calculate the Ahp tree}
\usage{
Calculate(ahpTree)
}
\arguments{
\item{ahpTree}{a data.tree tree containing the Ahp model specification}
}
\description{
Calculate the Ahp tree
}

