% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ahp_calculate.R
\name{CalculateAhp}
\alias{CalculateAhp}
\title{Calculate the Ahp tree}
\usage{
CalculateAhp(tr)
}
\arguments{
\item{tr}{a data.tree tree containing the Ahp model specification}
}
\description{
Calculate the Ahp tree
}

