% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ahp_functions.R
\name{CalculateAhpMatrix}
\alias{CalculateAhpMatrix}
\title{Calculate the ahp weights from the symmetric AHP matrix}
\usage{
CalculateAhpMatrix(mat, allowedConsistency = 1)
}
\arguments{
\item{mat}{The AHP preference matrix}

\item{allowedConsistency}{if the AHP consistency ratio is larger
than this value, AHP is not applied and equal weights are returned.}
}
\value{
the ahp preference weights
}
\description{
Calculate the ahp weights from the symmetric AHP matrix
}

