% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createLogicalPCM.R
\name{createLogicalPCM}
\alias{createLogicalPCM}
\title{Simulated Logical PCM For AHP}
\usage{
createLogicalPCM(ord, prefVec = rep(NA, ord))
}
\arguments{
\item{ord}{The order of the PCM desired}

\item{prefVec}{The preference vector of length = ord}
}
\value{
A Logical PCM
}
\description{
Creates a logical PCM for the AHP such as would be created by a rational decision maker based on a relative vector of preferences for the alternatives involved. Choices of the pairwise comparison ratios are from the Fundamental Scale and simulate a reasonable degree of error. The algorithm is modified from a paper by Bose, A [2022], DOI: https://doi.org/10.1002/mcda.1784
}
\examples{
lPCM <- createLogicalPCM(3,c(1,2,3));
lPCM <- createLogicalPCM(5,c(0.25,0.4,0.1,0.05,0.2));
}
