% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_combinations.R
\name{generate_combinations}
\alias{generate_combinations}
\title{Generate a set of criteria of preferences from AHP}
\usage{
generate_combinations(n, ...)
}
\arguments{
\item{n}{A number of defined criteria}

\item{...}{Arguments of combinations function}
}
\value{
generate_combinations returns an array of matrix object
}
\description{
Generate a set of criteria of preferences from AHP
}
\examples{
generate_combinations(3)
}
