% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ahp_calculate.R
\name{GetAlternativesNames}
\alias{GetAlternativesNames}
\title{Get the names of the alternatives in the tree}
\usage{
GetAlternativesNames(tr)
}
\arguments{
\item{tr}{The ahp tree}
}
\value{
a character vector containing the names of all the alternatives
}
\description{
Get the names of the alternatives in the tree
}

