% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analyze.R
\name{GetDataFrame}
\alias{GetDataFrame}
\title{Converts the tree into a data.frame, containing all the weights contributions
to the overall decision.}
\usage{
GetDataFrame(tr)
}
\arguments{
\item{tr}{the calculated AHP data.tree}
}
\value{
a data.frame containing the contribution of each criteria
}
\description{
Converts the tree into a data.frame, containing all the weights contributions
to the overall decision.
}

