% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{GetPairwiseFromFunction}
\alias{GetPairwiseFromFunction}
\title{Create table from function}
\usage{
GetPairwiseFromFunction(node)
}
\arguments{
\item{node}{The node}
}
\description{
Create table from function
}

