% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load.R
\name{LoadAhpFile}
\alias{LoadAhpFile}
\title{Load an ahp model from file}
\usage{
LoadAhpFile(file)
}
\arguments{
\item{file}{the path and name of the file to load}
}
\value{
a data.tree containing the model specification
}
\description{
Load an ahp model from file
}

