% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load.R
\name{LoadFile}
\alias{LoadFile}
\title{Load an ahp model from file}
\usage{
LoadFile(file)
}
\arguments{
\item{file}{The full path to the file to be loaded.}
}
\value{
A \code{\link{data.tree}} containing the model specification.
}
\description{
ahp files are in YAML format, and they are self-contained, fully specified ahp problems. They contain two sections: alternatives and goal.
The goal section is a tree of criteria, each criteria having a name, a preference attribute, and possible child criteria or alternatives.
}
\examples{
ahpFile <- system.file("extdata", "car.ahp", package="ahp")

#look at a sample file
cat(readChar(ahpFile, file.info(ahpFile)$size))

#load the file into R
carAhp <- LoadFile(ahpFile)
}

