% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ahp_functions.R
\name{PrioritiesFromPairwiseMatrixEigenvalues}
\alias{PrioritiesFromPairwiseMatrixEigenvalues}
\alias{PrioritiesFromPairwiseMatrixMeanNormalization}
\alias{PrioritiesFromPairwiseMatrixGeometricMean}
\title{Calculate the ahp priority weights from the AHP matrix.}
\usage{
PrioritiesFromPairwiseMatrixEigenvalues(mat, allowedConsistency = 1)

PrioritiesFromPairwiseMatrixMeanNormalization(mat, allowedConsistency = 1)

PrioritiesFromPairwiseMatrixGeometricMean(mat, allowedConsistency = 1)
}
\arguments{
\item{mat}{The AHP preference matrix}

\item{allowedConsistency}{if the AHP consistency ratio is larger 
than this value, AHP is not applied and equal weights are returned.}
}
\value{
the ahp preference weights
}
\description{
For a comparison of different methods, see for example \bold{How to derive priorities in AHP: a comparative study}, 
by Alessio Ishizaka and Markus Lusti, as available here: http://eprints.port.ac.uk/9041/1/filetodownload,70633,en.pdf
}
