% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui.R
\name{RunGUI}
\alias{RunGUI}
\title{Shows a GUI (Graphical User Interface) that lets you specify AHP models and view the results.}
\usage{
RunGUI()
}
\description{
Shows a GUI (Graphical User Interface) that lets you specify AHP models and view the results.
}

