% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/table.R
\name{ShowTable}
\alias{ShowTable}
\title{Displays the AHP analysis in form of an html table, with gradient
colors and nicely formatted.}
\usage{
ShowTable(tr, weightColor = "honeydew3", consistencyColor = "wheat2",
  alternativeColor = "thistle4")
}
\arguments{
\item{tr}{The calculated ahp tree}

\item{weightColor}{The name of the color to be used to emphasize weights of categories. See \code{color} for a list of possible colors.}

\item{consistencyColor}{The name of the color to be used to highlight bad consistency}

\item{alternativeColor}{The name of the color used to highlight big contributors to alternative choices.}
}
\value{
a \code{formattable} object which, in most environments, will be displayed as an HTML table
}
\description{
Displays the AHP analysis in form of an html table, with gradient
colors and nicely formatted.
}

