% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Testes.R
\name{tabela2_ahp_xlsx}
\alias{tabela2_ahp_xlsx}
\title{Tabela AHP para listas}
\usage{
tabela2_ahp_xlsx(caminho, mapeamento, alternativas = c())
}
\arguments{
\item{caminho}{caminho para um arquivo xlsx contendo as comporacoes entre criterios e alternativas, e por fim, comparacao entre criterios}

\item{mapeamento}{vetor contendo informações sobre as posições dos critérios e subcritérios.}

\item{alternativas}{vetor que contem o nome das alternativas, se vazio, sera gerado automaticamente 
pelas letras do alfabeto (A:Z)}
}
\value{
retorna tabela de relacoes de preferencia entre criterios
}
\description{
Essa funcao aplica o metodo AHP de saaty em listas
}
