% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/visualize.R
\name{Visualize}
\alias{GetGraph}
\alias{Visualize}
\title{Draw a diagram of the AHP tree.}
\usage{
Visualize(ahpTree, criteriaNodesStyle = list(style = "filled,rounded", shape =
  "box", color = "palevioletred"), alternativeNodesStyle = list(style =
  "rounded", shape = "box"), criteriaEdgesStyle = list(arrowhead = "vee"),
  alternativeEdgesStyle = list(dir = "none", color = "grey85", penwidth =
  0.5))

GetGraph(ahpTree, criteriaNodesStyle = list(style = "filled,rounded", shape =
  "box", color = "palevioletred"), alternativeNodesStyle = list(style =
  "rounded", shape = "box"), criteriaEdgesStyle = list(arrowhead = "vee"),
  alternativeEdgesStyle = list(dir = "none", color = "grey85", penwidth =
  0.5))
}
\arguments{
\item{ahpTree}{The tree to be drawn}

\item{criteriaNodesStyle}{a list of graphviz / dot language styling elements}

\item{alternativeNodesStyle}{a list of graphviz / dot language styling elements}

\item{criteriaEdgesStyle}{a list of graphviz / dot language styling elements}

\item{alternativeEdgesStyle}{a list of graphviz / dot language styling elements}
}
\description{
The function uses graphviz via DiagrammeR. For details on styling, refer
to these.
}

