% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ahp.package.R
\docType{package}
\name{ahp}
\alias{ahp}
\alias{ahp-package}
\title{ahp AHP (Analytic Hierarchy Process) Modeling for R}
\description{
AHP (Analytic Hierarchy Process) is a decision making framework developed by Thomas Saaty.
This package lets you model complex decision problems according to the AHP framework.
}

