% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AHP.R
\name{AHP}
\alias{AHP}
\title{Método AHP}
\usage{
AHP(caminho)
}
\arguments{
\item{caminho}{Um caminho para um arquivo xlsx que contenha as planilhas com os dados}
}
\value{
Retorna um tibble com as proporcoes para cada alternativa estudada.
}
\description{
Funcao que realiza o metodo AHP em um conjunto de dados apropriado.
}
\details{
Lembrar que na arrumacao das planilhas do excel devemos ter : 
MAC1, .... , MACn , MCC 
As planilhas devem ser nomeadas
A primeira coluna da mcc tem que ser a primeira planilha........
}
