% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analyze.R, R/table.R
\name{Analyze}
\alias{Analyze}
\alias{AnalyzeTable}
\title{Converts the tree into a data.frame, containing all the weights contributions
to the overall decision.}
\usage{
Analyze(ahpTree, decisionMaker = "Total", variable = c("weightContribution",
  "priority", "score"), sort = c("priority", "totalPriority", "orig"))

AnalyzeTable(ahpTree, decisionMaker = "Total",
  variable = c("weightContribution", "priority", "score"),
  sort = c("priority", "totalPriority", "orig"), weightColor = "honeydew3",
  consistencyColor = "wheat2", alternativeColor = "thistle4")
}
\arguments{
\item{ahpTree}{the calculated AHP \code{data.tree}}

\item{decisionMaker}{the name of the decision maker. The default returns the joint decision.}

\item{variable}{the variable to display, i.e. either weightContribution (the default), priority, or score}

\item{sort}{sort either by priority according to the decision maker (the default), by totalPriority, or as originally specified (orig)}

\item{weightColor}{The name of the color to be used to emphasize weights of categories. See \code{color} for a list of possible colors.}

\item{consistencyColor}{The name of the color to be used to highlight bad consistency}

\item{alternativeColor}{The name of the color used to highlight big contributors to alternative choices.}
}
\value{
a \code{data.frame} containing the contribution of each criteria

a \code{\link{formattable}} data.frame object which, in most environments, will be displayed as an HTML table
}
\description{
Converts the tree into a data.frame, containing all the weights contributions
to the overall decision.

Displays the AHP analysis in form of an html table, with gradient
colors and nicely formatted.
}

