% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{Calculate}
\alias{Calculate}
\title{Calculate the Ahp tree}
\usage{
Calculate(ahpTree, pairwiseFun = PrioritiesFromPairwiseMatrixEigenvalues,
  scoresFun = PrioritiesFromScoresDefault)
}
\arguments{
\item{ahpTree}{a data.tree tree containing the Ahp model specification}

\item{pairwiseFun}{lets you overwrite the function that is used to calculate the priorities from 
the pairwise preference matrix.}

\item{scoresFun}{lets you overwrite the function that is used to calculate the priorites from scores.}
}
\description{
Calculate the Ahp tree
}
