% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AHP.R
\docType{data}
\name{dados1}
\alias{dados1}
\title{Exemplo de dados 1}
\format{Exemplo de dados obtidos para utlizar o método AHP}
\source{
dados1
}
\usage{
dados1
}
\description{
Exemplo de dados 1
}
\keyword{datasets}
